{"schema":"decodex/commit/1","summary":"split connection and evidence rendering modules","authority":"manual"}
